Implemented gradient, I think Added neohookean gradient, created dF_dU function based on CSC417 A3 Fixed bug in deformation_gradient Changed T to be int matrix

+++ include/deformation_gradient.h
@@ -1,8 +1,15 @@
 #include <Eigen/Core>
 
+// Inputs:
+//   V    #V by 3 list of rest (initial pose) mesh vertex positions
+//   tet  list of vertices of tetrahedron
+//   Ur   #V by 3 rig displacement at the current frame
+//   Uc   #V by 3 complementary displacement at the current frame
+// Outputs:
+//   F    3 by 3 deformation gradient of tetrahedron
 void deformation_gradient(
   const Eigen::MatrixXd & V,
-  const Eigen::RowVector4d & tet,
+  const Eigen::RowVector4i & tet,
   const Eigen::VectorXd & Ur,
   const Eigen::VectorXd & Uc,
   Eigen::Matrix3d & F);
